requirements.txt
APScheduler==3.11.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
python-dotenv==1.1.0
requests==2.32.3
tzlocal==5.3.1
urllib3==2.4.0
